feat: scratch docker image
This moves to a scratch based docker image to reduce image size and
attack surface further.

The new image uses pyinstaller and staticx to compile the python app to
a binary and staticly link it. This binary is copied into a scratch
container and set to run as a non root user