fix(branch_status_checks): Fixed Branch Status check changes
The underlying python API is stale; it does a type of REST API call that is not supported.  I also fixed some issues with false positives when the order of items in a list is not equal.  Additionally, found and fixed some errors where the config is null and the repo is not or vice versa.